backend/cloud_functions/remove-bg/requirements.txt
boto3==1.34.5
fastapi==0.105.0
functions-framework==3.*
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
requests==2.31.0
SQLAlchemy==1.4.50
transparent-background==1.2.9
